+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * puddelgaming.com to reuploader
 * debug command to enable debugging
 * personal command for DasMaxiking
+* teamtrees command
 
 ### Changed